+++ console/src/components/Common/NotSupportedNote/index.tsx
@@ -5,6 +5,7 @@ const driverToLabel: Record<Driver, string> = {
   mysql: 'MySQL',
   postgres: 'PostgreSQL',
   mssql: 'MS Server',
+  bigquery: 'Big Query',
 };
 
 type NotSupportedNoteProps = {