+++ console/src/components/Services/Data/Add/AddExistingTableViewActions.js
@@ -8,8 +8,12 @@ import {
   getTableModifyRoute,
   getFunctionModifyRoute,
 } from '../../../Common/utils/routesUtils';
-import { dataSource } from '../../../../dataSources';
-import { findTable, escapeTableColumns } from '../../../../dataSources/common';
+import { dataSource, currentDriver } from '../../../../dataSources';
+import {
+  findTable,
+  escapeTableColumns,
+  getQualifiedTableDef,
+} from '../../../../dataSources/common';
 import { exportMetadata } from '../../../../metadata/actions';
 import {
   getUntrackTableQuery,
@@ -39,21 +43,24 @@ const addExistingTableSql = (name, customSchema, skipRouting = false) => {
       : getState().tables.currentSchema;
     const currentDataSource = getState().tables.currentDataSource;
     const tableName = name ? name : state.tableName.trim();
-    const tableDef = { name: tableName, schema: currentSchema };
+
+    const tableDef = getQualifiedTableDef(
+      {
+        name: tableName,
+        schema: currentSchema,
+      },
+      currentDriver
+    );
+
     const table = findTable(getState().tables.allSchemas, tableDef);
+
     const requestBodyUp = getTrackTableQuery({
       tableDef,
       source: currentDataSource,
       customColumnNames: escapeTableColumns(table),
     });
 
-    const requestBodyDown = getUntrackTableQuery(
-      {
-        name: tableName,
-        schema: currentSchema,
-      },
-      currentDataSource
-    );
+    const requestBodyDown = getUntrackTableQuery(tableDef, currentDataSource);
 
     const migrationName = `add_existing_table_or_view_${currentSchema}_${tableName}`;
 
@@ -67,19 +74,20 @@ const addExistingTableSql = (name, customSchema, skipRouting = false) => {
           t => t.table_name === tableName && t.table_schema === currentSchema
         );
         const isTableType = dataSource.isTable(newTable);
-        const nextRoute = isTableType
-          ? getTableModifyRoute(
-              currentSchema,
-              currentDataSource,
-              tableName,
-              isTableType
-            )
-          : getTableBrowseRoute(
-              currentSchema,
-              currentDataSource,
-              tableName,
-              isTableType
-            );
+        const nextRoute =
+          isTableType && currentDriver !== 'bigquery'
+            ? getTableModifyRoute(
+                currentSchema,
+                currentDataSource,
+                tableName,
+                isTableType
+              )
+            : getTableBrowseRoute(
+                currentSchema,
+                currentDataSource,
+                tableName,
+                isTableType
+              );
         if (!skipRouting) {
           dispatch(_push(nextRoute));
         }
@@ -178,12 +186,17 @@ const addAllUntrackedTablesSql = tableList => {
     dispatch(showSuccessNotification('Adding...'));
     const bulkQueryUp = [];
     const bulkQueryDown = [];
+
     for (let i = 0; i < tableList.length; i++) {
       if (tableList[i].table_name !== 'schema_migrations') {
-        const tableDef = {
-          name: tableList[i].table_name,
-          schema: currentSchema,
-        };
+        const tableDef = getQualifiedTableDef(
+          {
+            name: tableList[i].table_name,
+            schema: currentSchema,
+          },
+          currentDriver
+        );
+
         const table = findTable(getState().tables.allSchemas, tableDef);
         bulkQueryUp.push(
           getTrackTableQuery({
@@ -197,7 +210,9 @@ const addAllUntrackedTablesSql = tableList => {
             {
               table: {
                 name: tableList[i].table_name,
-                schema: currentSchema,
+                [currentDriver === 'bigquery'
+                  ? 'dataset'
+                  : 'schema']: currentSchema,
               },
             },
             currentDataSource
@@ -205,6 +220,7 @@ const addAllUntrackedTablesSql = tableList => {
         );
       }
     }
+
     const migrationName = 'add_all_existing_table_or_view_' + currentSchema;
 
     const requestMsg = 'Adding existing table/view...';