+++ console/src/components/Services/Data/DataActions.js
@@ -26,7 +26,11 @@ import {
   cascadeUpQueries,
   getDependencyError,
 } from './utils';
-import { mergeDataMssql, mergeLoadSchemaDataPostgres } from './mergeData';
+import {
+  mergeDataMssql,
+  mergeLoadSchemaDataPostgres,
+  mergeDataBigQuery,
+} from './mergeData';
 import _push from './push';
 import { convertArrayToJson } from './TableModify/utils';
 import { CLI_CONSOLE_MODE, SERVER_CONSOLE_MODE } from '../../../constants';
@@ -146,7 +150,9 @@ const loadSchema = configOptions => {
         (!configOptions.tables || configOptions.tables.length === 0))
     ) {
       configOptions = {
-        schemas: [getState().tables.currentSchema],
+        schemas: [
+          getState().tables.currentSchema || getState().tables.schemaList[0],
+        ],
       };
     }
 
@@ -171,7 +177,6 @@ const loadSchema = configOptions => {
         );
       }
     }
-
     const body = {
       type: 'bulk',
       source,
@@ -224,6 +229,9 @@ const loadSchema = configOptions => {
               break;
             case 'mssql':
               mergedData = mergeDataMssql(data, metadataTables);
+              break;
+            case 'bigquery':
+              mergedData = mergeDataBigQuery(data, metadataTables);
               break;
             default:
           }
@@ -308,9 +316,14 @@ const setConsistentSchema = data => ({
 const fetchDataInit = (source, driver) => (dispatch, getState) => {
   const url = Endpoints.query;
 
-  const { schemaFilter } = getState().tables;
-  const currentSource = source || getState().tables.currentDataSource;
+  let { schemaFilter } = getState().tables;
 
+  if (driver === 'bigquery')
+    schemaFilter = getState().metadata.metadataObject.sources.find(
+      x => x.name === source
+    ).configuration.datasets;
+
+  const currentSource = source || getState().tables.currentDataSource;
   const query = getRunSqlQuery(
     dataSource.schemaListSql(schemaFilter),
     currentSource,
@@ -340,7 +353,6 @@ const fetchDataInit = (source, driver) => (dispatch, getState) => {
         type: FETCH_SCHEMA_LIST,
         schemaList,
       });
-
       let newSchema = '';
       if (schemaList.length) {
         newSchema =
@@ -350,7 +362,6 @@ const fetchDataInit = (source, driver) => (dispatch, getState) => {
             : schemaList.sort(Intl.Collator().compare)[0];
       }
       dispatch({ type: UPDATE_CURRENT_SCHEMA, currentSchema: newSchema });
-
       return dispatch(updateSchemaInfo()); // TODO
     },
     error => {
@@ -558,7 +569,6 @@ export const getDatabaseTableTypeInfo = (
       resolve({});
     });
   }
-
   const url = Endpoints.query;
   const sql = services[sourceType].getTableInfo(tables);
   const query = getRunSqlQuery(sql, sourceName, false, false, sourceType);
@@ -582,13 +592,18 @@ export const getDatabaseTableTypeInfo = (
       try {
         if (currentDriver === 'mssql') {
           res = JSON.parse(result.slice(1).join());
+        } else if (currentDriver === 'bigquery') {
+          res = result.slice(1).map(t => ({
+            table_name: t[0],
+            table_schema: t[1],
+            table_type: t[2],
+          }));
         } else {
           res = JSON.parse(result[1]);
         }
-      } catch {
+      } catch (err) {
         res = [];
       }
-
       res.forEach(i => {
         if (
           !trackedTables.some(