+++ console/src/metadata/sourcesUtils.ts
@@ -11,6 +11,10 @@ export const addSource = (
       idle_timeout?: number;
       retries?: number;
     };
+    bigQuery: {
+      projectId: string;
+      datasets: string;
+    };
   },
   // supported only for PG sources at the moment
   replicas?: Omit<SourceConnectionInfo, 'connection_string'>[]
@@ -25,6 +29,20 @@ export const addSource = (
             connection_string: payload.dbUrl,
             pool_settings: payload.connection_pool_settings,
           },
+        },
+      },
+    };
+  }
+
+  if (driver === 'bigquery') {
+    return {
+      type: 'bigquery_add_source',
+      args: {
+        name: payload.name,
+        configuration: {
+          service_account: payload.dbUrl,
+          project_id: payload.bigQuery.projectId,
+          datasets: payload.bigQuery.datasets.split(',').map(d => d.trim()),
         },
       },
     };
@@ -53,6 +71,9 @@ export const removeSource = (driver: Driver, name: string) => {
       break;
     case 'mysql':
       prefix = 'mysql_';
+      break;
+    case 'bigquery':
+      prefix = 'bigquery_';
       break;
     default:
       prefix = 'pg_';